[SPARK-44535][CONNECT][SQL] Move required Streaming API to sql/api
### What changes were proposed in this pull request?
This PR moves a bunch streaming classed to the SQL/API project.

### Why are the changes needed?
This is needed to disconnect the Spark Connect Scala Client from catalyst.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing Tests.

Closes #42140 from hvanhovell/SPARK-44535.

Authored-by: Herman van Hovell <herman@databricks.com>
